# mnemosyne-web-interface/mnemosyne-api/docs/swagger-spec.yaml
openapi: 3.0.0
paths:
  /api/users/forgot-password:
    post:
      operationId: UsersController_forgotPassword
      summary: Endpoint is responsible for initiation of password reset.
      parameters: []
      requestBody:
        required: true
        description: In order to reset password user has to provide only email.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordDto"
      responses:
        "201":
          description: In both cases - if user exists or not, user will get the same
            response saying that email with password reset link has been sent.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResetPasswordEmailDto"
        "400":
          description: Bad request error is sent to user if user will try to send more
            than 1 reset request within 180 seconds.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WrongTimeframeException"
        "403":
          description: In case if password has been changed within 24h, user will not be
            able to continue.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasswordChangedException"
      tags: &a1
        - Users
      security:
        - basicAuth: []
  /api/users/upload-user-photo:
    post:
      operationId: UsersController_uploadUserPhoto
      summary: Endpoint is responsible for upload of the user photo.
      parameters: []
      requestBody:
        required: true
        description: Body contains base64-encoded PNG picture.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadPhotoDto"
      responses:
        "201":
          description: As a response user gets message that photo is uploaded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhotoUploadedDto"
        "400":
          description: If user tries to upload something else except of base64-encoded PNG
            picture.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WrongPictureException"
      tags: *a1
      security:
        - x-access-token: []
        - basicAuth: []
  /api/users/user-info:
    get:
      operationId: UsersController_getUserInfo
      summary: Endpoint is responsible for getting user personal information.
      parameters: []
      responses:
        "201":
          description: As a response user gets personal information.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetUserInfoResponseDto"
      tags: *a1
      security:
        - x-access-token: []
        - basicAuth: []
    patch:
      operationId: UsersController_updateUserInfo
      summary: Endpoint is responsible for updating user information.
      parameters: []
      requestBody:
        required: true
        description: Body contains user personal information.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserInfoDto"
      responses:
        "201":
          description: As a response user gets message that information has been updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserUpdatedDto"
      tags: *a1
      security:
        - x-access-token: []
        - basicAuth: []
  /api/users/user-security:
    get:
      operationId: UsersController_getUserSecuritySettings
      summary: Endpoint is responsible for getting user security settings.
      parameters: []
      responses:
        "201":
          description: As a response user gets security settings.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetUserSecResponseDto"
      tags: *a1
      security:
        - x-access-token: []
        - basicAuth: []
  /api/users/marketplace-user:
    get:
      operationId: UsersController_getMarketplaceUserById
      summary: Endpoint is responsible for getting public marketplace user information.
      parameters:
        - name: marketplaceUserId
          required: true
          in: query
          description: Marketplace user ID (or just user ID)
          schema:
            type: string
      responses:
        "201":
          description: As a response user gets a message with public information about
            user along with optional information about company.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetMarketplaceUserByIdDto"
        "404":
          description: In case if marketplace user ID has been modified and user has not
            been found user gets not found error message.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserNotFoundException"
      tags: *a1
      security:
        - basicAuth: []
...
info:
  title: Mnemosyne
  description: Documentation of Mnemosyne API.
  version: 0.0.1
  contact: {}
...
servers: []
components:
  securitySchemes:
    x-access-token:
      scheme: bearer
      bearerFormat: JWT
      description: "Please enter the JWT token in the following format: Bearer <JWT>"
      name: x-access-token
      type: apiKey
      in: header
    basicAuth:
      type: http
      scheme: basic
      description: Basic API authentication username and password
      in: header
  schemas:
    ResetPasswordEmailDto:
      type: object
      properties:
        message:
          type: string
          description: Reset password email sent response message
          example: reset-password-email-sent
      required:
        - message
    ForgotPasswordDto:
      type: object
      properties:
        email:
          type: string
          description: Email address
          example: joe@doe.com
        language:
          type: string
          enum:
            - ru
            - pl
            - en
          description: One of 3 available languages
          example: pl
      required:
        - email
    PasswordChangedException:
      type: object
      properties:
        message:
          type: string
          description: Password already changed error message
          example: password-already-changed
      required:
        - message
    WrongTimeframeException:
      type: object
      properties:
        message:
          type: string
          description: Wrong timeframe error message
          example: wrong-timeframe
      required:
        - message
    PhotoUploadedDto:
      type: object
      properties:
        message:
          type: string
          description: Photo uploaded response message
          example: photo-uploaded
      required:
        - message
    UploadPhotoDto:
      type: object
      properties:
        userPhoto:
          type: string
          description: User base64-encoded images
          example: data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAUAAAAFCAYAAACNbyblAAAAHElEQVQI12P4...
      required:
        - userPhoto
    WrongPictureException:
      type: object
      properties:
        message:
          type: string
          description: Wrong picture format error message
          example: wrong-picture-format
      required:
        - message
    GetUserInfoResponseDto:
      type: object
      properties:
        userIdHash:
          type: string
          description: User MD5 hashed ID
          example: 71956c9eea017cb85242a49f416fd859
        firstName:
          type: string
          description: First name
          example: Joe
        lastName:
          type: string
          description: Last name
          example: Doe
        email:
          type: string
          description: Email address
          example: joe@doe.com
        isProfilePicPresent:
          type: boolean
          description: Is user profile picture present
          example: "true"
        namePronunciation:
          type: string
          description: This could be a phonetic pronunciation, or an example of something
            your name sounds like
          example: Zoe ("zo-ee")
        homeAddress:
          type: string
          description: User home address
          example: One Apple Park Way, United States
        homePhone:
          type: string
          description: User home phone
          example: (800) 1231-1234
        isCompanyMember:
          type: boolean
          description: Flag if the user is company member
          example: "false"
        companyName:
          type: string
          description: Company name
          example: Apple Inc.
        companyId:
          type: string
          description: Company ID
          example: 4eea51f7-ebf1-4470-b54e-eb3b8c79bf2c
        roleName:
          type: string
          description: Company custom role name
          example: My Test Role
        roleScopes:
          description: The array of the scopes for company custom role
          example:
            - USER_MANAGEMENT
          type: array
          items:
            type: array
      required:
        - userIdHash
        - firstName
        - lastName
        - email
        - isProfilePicPresent
        - namePronunciation
        - homeAddress
        - homePhone
        - isCompanyMember
        - companyName
        - companyId
        - roleName
        - roleScopes
    PhoneStatus:
      type: object
      properties:
        isSetUp:
          type: boolean
          description: Phone setup - is phone set up
          example: "true"
        twoLastDigit:
          type: string
          description: Phone setup - 2 last digits
          example: "12"
      required:
        - isSetUp
        - twoLastDigit
    GetUserSecResponseDto:
      type: object
      properties:
        phoneStatus:
          $ref: "#/components/schemas/PhoneStatus"
        passwordCanBeChanged:
          type: boolean
          description: Password can be changed
          example: "true"
        emailChanged:
          type: boolean
          description: Is email changed
          example: "true"
        isTwoFaSetUp:
          type: boolean
          description: Is 2FA set up
          example: "true"
        email:
          type: string
          description: Email address
          example: joe@doe.com
      required:
        - phoneStatus
        - passwordCanBeChanged
        - emailChanged
        - isTwoFaSetUp
        - email
    UserUpdatedDto:
      type: object
      properties:
        message:
          type: string
          description: User updated response message
          example: user-updated
      required:
        - message
    UpdateUserInfoDto:
      type: object
      properties:
        firstName:
          type: string
          description: First name
          example: Joe
        lastName:
          type: string
          description: Last name
          example: Doe
        namePronunciation:
          type: string
          description: This could be a phonetic pronunciation, or an example of something
            your name sounds like
          example: Zoe ("zo-ee")
        homeAddress:
          type: string
          description: User home address
          example: One Apple Park Way, United States
        homePhone:
          type: string
          description: User home phone
          example: (800) 1231-1234
    UserNotFoundException:
      type: object
      properties:
        message:
          type: string
          description: Marketplace user not found error message
          example: user-not-found
      required:
        - message
    MarketplaceUser:
      type: object
      properties:
        email:
          type: string
          description: Email address
          example: joe@doe.com
        firstName:
          type: string
          description: First name
          example: Joe
        lastName:
          type: string
          description: Last name
          example: Doe
        homeAddress:
          type: string
          description: User home address
          example: One Apple Park Way, United States
        homePhone:
          type: string
          description: User home phone
          example: (800) 1231-1234
        namePronunciation:
          type: string
          description: This could be a phonetic pronunciation, or an example of something
            your name sounds like
          example: Zoe ("zo-ee")
        createdAt:
          format: date-time
          type: string
          description: User creation date
          example: 2023-11-17T11:42:26.534Z
        userIdHash:
          type: string
          description: User MD5 hashed ID
          example: 71956c9eea017cb85242a49f416fd859
        companyId:
          type: string
          description: Company ID
          example: 4eea51f7-ebf1-4470-b54e-eb3b8c79bf2c
        companyName:
          type: string
          description: Company name
          example: Apple Inc.
      required:
        - email
        - firstName
        - lastName
        - homeAddress
        - homePhone
        - namePronunciation
        - createdAt
        - userIdHash
        - companyId
        - companyName
    GetMarketplaceUserByIdDto:
      type: object
      properties:
        marketplaceUser:
          description: Public information of the user on the marketplace
          example:
            email: joe@doe.com
            firstName: Joe
            lastName: Doe
            homeAddress: One Apple Park Way, United States
            homePhone: (800) 1231-1234
            namePronunciation: Zoe ("zo-ee")
            createdAt: 2023-11-17T11:42:26.534Z
            userIdHash: 71956c9eea017cb85242a49f416fd859
            companyId: 4eea51f7-ebf1-4470-b54e-eb3b8c79bf2c
            companyName: Apple Inc.
          allOf:
            - $ref: "#/components/schemas/MarketplaceUser"
      required:
        - marketplaceUser
...