# mnemosyne-web-interface/mnemosyne-front/src/libs/api/responses/users/get-marketplace-user-by-id.interface.ts
export interface MarketplaceUser {
  email: string | undefined;
  firstName: string;
  lastName: string;
  homeAddress: string | null;
  homePhone: string | null;
  namePronunciation: string | null;
  createdAt: Date;
  userIdHash: string | null;
  companyId: string | null;
  companyName: string | null;
}

export interface GetMarketplaceUserByIdResponse {
  marketplaceUser: MarketplaceUser;
}